Create a fixture for the SnapshotService
